CALCULATOR BUTTON LAYOUT
Calculator buttons laid out by changing order in the html. Also
by removing the border in the CSS.

Small changes in script.js due to changing the text content of two
of the buttons in the html.

Other changes made in the CSS may not be permanent as colour scheme
hasn't been entirely settled on.