ADD ERROR MESAGE FOR DISPLAY OVERFLOW SOLVE DECIMAL POINT ISSUE ON INPUT AFTER ANSWER
Added an error message for when the answer or user inputs overflow on the
display element. Also set overflow to hidden on the display element.

Found a small issue with not being able to use the answer for a second
calculation if the first input was the decimal point. This was resolved
by setting num1 to equal answer then answer, num2, and operator back to
empty strings in the calculation function.